fix: app init missing logger, globalHandler and intl-pluralrules polyfill
